requirements.txt
python-telegram-bot[webhooks]==20.6
python-dotenv
httpx~=0.25.0
